Implement -b / --buffer-size
This adds -b / --buffer-size to the peco binary. It basically limits the
maximum number of lines that peco will hold at any given time, allowing
you to keep peco running for a long time, even when you are receiving an
infinite stream of lines.

By default the buffer size is unlimited. If --buffer-size is specified,
lines will be thrown away in FIFO order