Draw prompt before running the query
This basically makes it easier for the users to see what exactly they
are waiting for. Previously we ran the filter, then re-drew the screen.
But this could take a while if the buffer is really big or the matcher
is slow, and therefore you would be stuck with a really old version of
your query in the screen for a while.

This change basically forces a redraw of the prompt alone before running
the query, which was something I meant to do for a long time, and now
possible because I split up the drawing in the layout object to
Prompt/Status/Lines sections