Remove InputReader
Upon realizing that I had written this but haven't used it in actual
peco code other than testing, I tried it: It didn't work. Well, at least
it didn't work out of the box.

I could sense a loooooong debugging session ahead, and decided it's not
worth it at this point in time. we can always re-instate it, as we have
the history of it in git