Tweak timing for when to initialize the screen and other components
We should wait to initialize the screen until we have something
coming in from the source: otherwise multiple peograms chained via
pipes would compete for the same terminal resource, and things may
go southward.

But then there are components that rely on the screen being intialized,
so we let them wait as well.